test(mlx): consolidate gated-delta tests into one file
Merges the binding-sweep/detection suite and the gradient-parity suite
into tests/test_mlx_gated_delta_vjp.py. The torch shim now installs only
when real mlx is absent (CI), parity tests skip without Metal, and the
module-restore fixture keeps the file order-independent against
shim-based sibling files. Folds source-patched/foreign-untouched asserts
into the sweep test, merges detection cases, and parametrizes parity
over both VJP implementations.
